TT2 template and HTML output for cheatsheet
Text was originally extracted from the source modules
via Pod::POM, had to do a lot of massaging to get it all
to fit so I'll try to put together a script for automating
those changes so it's easier to keep up to date in future.

Actual PDF generation was done via Firefox, since that
seems to give results closest to the original styles,
but ideally would replace this with something automated
like wkhtmltopdf.